requirements.txt
fastapi
uvicorn
google-adk
google-genai
google-cloud-aiplatform
google-cloud-logging
python-dotenv